requirements.txt
alembic==1.13.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.5
et-xmlfile==1.1.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flask-mail==0.9.1
flask-restful==0.3.10
Flask-Caching==1.10.1
greenlet==3.1.1
gunicorn==20.1.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
mysql-connector-python==9.0.0
mysqlclient==2.1.0
PyMySQL==1.1.1
nest-asyncio==1.6.0
numpy==1.21.2
openpyxl==3.1.2
packaging==24.1
pandas==1.3.3
plotly==5.24.1
pycparser==2.22
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2024.2
reportlab==4.0.4
requests==2.32.3
retrying==1.3.4
six==1.16.0
SQLAlchemy==1.4.23
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.17.6
Werkzeug==2.0.1
WTForms==3.1.2
zipp==3.20.2
zope.interface==7.1.1
fastapi
pytest==7.4.0
coverage==7.3.0
# Remove psycopg2 since we're using MySQL